[feat/TK-141]: 공동결제내역 충돌 방지
