--- src/main/java/withbeetravel/service/SharedPaymentServiceImpl.java
@@ -1,169 +1,0 @@
-package withbeetravel.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
-import withbeetravel.domain.SharedPayment;
-import withbeetravel.domain.Travel;
-import withbeetravel.dto.response.SharedPaymentRecordResponse;
-import withbeetravel.dto.response.SharedPaymentResponse;
-import withbeetravel.dto.response.SuccessResponse;
-import withbeetravel.exception.CustomException;
-import withbeetravel.exception.error.PaymentErrorCode;
-import withbeetravel.exception.error.TravelErrorCode;
-import withbeetravel.exception.error.ValidationErrorCode;
-import withbeetravel.repository.SharedPaymentRepository;
-import withbeetravel.repository.TravelMemberRepository;
-import withbeetravel.repository.TravelRepository;
-
-import java.io.IOException;
-import java.time.LocalDate;
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Service
-@RequiredArgsConstructor
-public class SharedPaymentServiceImpl implements SharedPaymentService{
-
-    private final S3Uploader s3Uploader;
-
-    private final TravelRepository travelRepository;
-    private final SharedPaymentRepository sharedPaymentRepository;
-    private final TravelMemberRepository travelMemberRepository;
-
-    // S3에 이미지를 저장할 경로
-    private static final String SHARED_PAYMENT_IMAGE_DIR = "shared-payments";
-
-    @Override
-    @Transactional
-    public SuccessResponse addAndUpdatePaymentRecord(
-            Long travelId,
-            Long sharedPaymentId,
-            MultipartFile image,
-            String comment,
-            boolean isMainImage) {
-
-        // SharedPayment Entity 가져오기
-        SharedPayment sharedPayment = sharedPaymentRepository.findById(sharedPaymentId)
-                .orElseThrow(() -> new CustomException(PaymentErrorCode.SHARED_PAYMENT_NOT_FOUND));
-
-        // 이미지 추가, 수정, 삭제
-        String newImage = uploadImage(travelId, sharedPayment, image);
-
-        // 새로운 image 정보 엔티티에서 변경
-        sharedPayment.updatePaymentImage(newImage);
-
-        // comment 정보 엔티티에서 변경
-        sharedPayment.updatePaymentCommnet(comment);
-
-        // 메인 이미지로 설정해준 경우 여행 정보 수정
-        if(isMainImage && image != null && newImage != null) {
-            // 여행 정보 찾아오기
-            Travel travel = travelRepository.findById(travelId)
-                    .orElseThrow(() -> new CustomException(TravelErrorCode.TRAVEL_NOT_FOUND));
-
-            // 여행 이미지 수정
-            travel.updateMainImage(newImage);
-        }
-
-        return SuccessResponse.of(200, "이미지 및 문구를 성공적으로 변경하였습니다.");
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public SuccessResponse<SharedPaymentRecordResponse> getSharedPaymentRecord(Long sharedPaymentId) {
-
-        // SharedPayment 엔티티 가져오기
-        SharedPayment sharedPayment = sharedPaymentRepository.findById(sharedPaymentId)
-                .orElseThrow(() -> new CustomException(PaymentErrorCode.SHARED_PAYMENT_NOT_FOUND));
-
-        // Response Dto에 담기
-        SharedPaymentRecordResponse responseDto = SharedPaymentRecordResponse.from(sharedPayment);
-
-        return SuccessResponse.of(200, "여행 기록 불러오기 성공", responseDto);
-    }
-
-    @Override
-    @Transactional(readOnly = true)
-    public SuccessResponse<Page<SharedPaymentResponse>> getSharedPaymentAll(Long travelId,
-                                                                            int page,
-                                                                            String sortBy,
-                                                                            Long memberId,
-                                                                            LocalDate startDate,
-                                                                            LocalDate endDate) {
-        // 정렬 타입 검증
-        if (!sortBy.equals("latest") && !sortBy.equals("amount")) {
-            throw new CustomException(PaymentErrorCode.INVALID_SORT_TYPE);
-        }
-
-        // 날짜 범위 검증
-        if (startDate != null && endDate != null && startDate.isAfter(endDate)) {
-            throw new CustomException(ValidationErrorCode.DATE_RANGE_ERROR);
-        }
-
-        // 멤버 ID가 제공된 경우, 해당 멤버가 이 여행의 멤버인지 확인
-        if (memberId != null) {
-            boolean isMemberOfTravel = travelMemberRepository.existsByTravelIdAndId(travelId, memberId);
-            if (!isMemberOfTravel) {
-                throw new CustomException(PaymentErrorCode.NON_TRAVEL_MEMBER_INCLUDED);
-            }
-        }
-
-        // sortBy가 amount면 금액 내림차순, 아니면 날짜 내림차순
-        Pageable pageable = PageRequest.of(page, 10,
-                Sort.by(Sort.Direction.DESC, sortBy.equals("amount") ? "paymentAmount" : "paymentDate"));
-
-        // 해당 여행의 공동 결제 내역 페이지 조회
-        Page<SharedPayment> sharedPayments = sharedPaymentRepository.findAllByTravelIdAndMemberIdAndDateRange(
-                travelId, memberId, startDate, endDate, pageable);
-
-        if (sharedPayments.isEmpty()) {
-            throw new CustomException(PaymentErrorCode.SHARED_PAYMENT_NOT_FOUND);
-        }
-
-        // Response DTO 변환
-        Page<SharedPaymentResponse> responseDtos = SharedPaymentResponse.from(sharedPayments);
-
-        return SuccessResponse.of(200, "모든 공동 결제 내역 조회 성공", responseDtos);
-    }
-
-    // 이미지 추가, 수정, 삭제
-    private String uploadImage(Long travelId, SharedPayment sharedPayment, MultipartFile image) {
-
-        // 원래 이미지
-        String paymentImage = sharedPayment.getPaymentImage();
-
-        // 새로 추가한 이미지
-        String newImage = null;
-
-        // image가 새로 들어왔다면 S3에 저장
-        if(image != null) {
-
-            // 이미지 저장할 S3 디렉토리 정보
-            String dirName = SHARED_PAYMENT_IMAGE_DIR + "/" + travelId;
-
-            try {
-                if(paymentImage != null) { // 해당 공동결제 내역에 이미 이미지가 있다면 업데이트
-                    newImage = s3Uploader.update(image, paymentImage, dirName);
-                } else { // 없다면 업로드
-                    newImage = s3Uploader.upload(image, dirName);
-                }
-            } catch (IOException e) { // 이미지 저장에 실패했을 경우
-                throw new CustomException(ValidationErrorCode.IMAGE_PROCESSING_FAILED);
-            }
-        }
-        // image가 null로 들어왔다면, 기존 이미지 삭제
-        else {
-            // S3에 이미지 삭제
-            s3Uploader.delete(paymentImage);
-        }
-
-        // 새로운 이미지 정보 반환
-        return newImage;
-    }
-}
